requirements.txt
# Python dependencies cho Secure Assignment System
Flask==3.1.1
Werkzeug==3.1.3
pycryptodome==3.23.0
click==8.2.1
flask-moment==1.0.6
colorama==0.4.6
tzdata==2025.2
requests==2.32.4

# ========================================
# HUONG DAN CAI DAT & CHAY DU AN TOAN DIEN
# ========================================

# 1. Cai Python 3.8+ tai https://www.python.org/downloads/
#    Kiem tra: python --version

# 2. Tao & kich hoat moi truong ao (venv):
#    - Windows:
    #    python -m venv venv
    #    venv\Scripts\activate
#    - Mac/Linux:
#        python3 -m venv venv
#        source venv/bin/activate

# 3. Cai dependencies Python:
#        pip install -r requirements.txt

# 4. Cai Node.js & npm:
#    - Windows: winget install OpenJS.NodeJS
#    - Mac: brew install node
#    - Linux: sudo apt install nodejs npm
#    Kiem tra: node --version, npm --version

# 5. Cai LocalTunnel:
#        npm install -g localtunnel
#        lt --version

# 6. Chay web app:
#        python web_app.py

# 7. Chia se web ra ngoai:
#        lt --port 5000 --subdomain myassignment
#    => Nhan link: https://myassignment.loca.lt

# 8. Neu bi hoi Tunnel Password:
#    - Lay IP public tai https://www.whatismyip.com/
#    - Nhap vao o Tunnel Password

# 9. Xem/sua database:
#    - Dung DB Browser for SQLite hoac extension SQLite tren VSCode
#    - Mo file data/assignment_system.db

# 10. Tat he thong:
#    - Dung web app: Ctrl+C
#    - Dung LocalTunnel: Ctrl+C

# ========================================
# Note: Luon kich hoat venv truoc khi chay lenh Python!
# ========================================

# Thao tac cai nhanh
# - Windows (CMD/PowerShell):
# python -m venv venv
# venv\Scripts\activate
# pip install -r requirements.txt
# python web_app.py

# winget install OpenJS.NodeJS
# npm install -g localtunnel
# lt --port 5000 --subdomain hatuananh

# - Mac/Linux (Bash):
# python3 -m venv venv
# source venv/bin/activate
# pip install -r requirements.txt
# python3 web_app.py

# brew install node    # (hoac sudo apt install nodejs npm tren Ubuntu)
# npm install -g localtunnel
# lt --port 5000 --subdomain hatuananh



# # ========================================

# ````
# ## Cac lenh SQL co ban

# Duoi day la mot so lenh huu ich de ban thao tac voi database. Hay copy, sua va chay chung trong Query Editor.

# ---

# ### 1. Xem du lieu (An toan)

# **Xem tat ca thong tin cua 10 file gan nhat:**
# ```sql
# SELECT * FROM files ORDER BY id DESC LIMIT 10;
# ````

# **Chi xem ten file va trang thai cua cac file dang xu ly (`pending`):**

# ```sql
# SELECT id, filename, status FROM files WHERE status = 'pending';
# ```

# **Dem tong so file:**

# ```sql
# SELECT COUNT(*) FROM files;
# ```

# ### 2. Sua du lieu (Can than!)

# **Doi trang thai cua file co ID cu the thanh `completed`:**

# ```sql
# UPDATE files
# SET status = 'completed'
# WHERE id = 20; -- Thay 20 bang ID ban muon
# ```

# **Sua ten file cua mot ID cu the:**

# ```sql
# UPDATE files
# SET filename = 'ten_file_moi.zip'
# WHERE id = 3; -- Thay 3 bang ID ban muon
# ```

# ### 3. Xoa du lieu (Rat nguy hiem!)

# **Xoa mot file voi ID cu the:**

# ```sql
# DELETE FROM files
# WHERE id = 4; -- Thay 4 bang ID ban muon xoa
# ```

# **Xoa TAT CA cac file co trang thai `error`:**

# ```sql
# DELETE FROM files
# WHERE status = 'error';
# ```